# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Ones API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      operationId: health
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /v1/events/covers/generate:
    post:
      operationId: generateEventCover
      summary: Generate an AI event cover preview and return a pre-signed URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateEventCoverRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateEventCoverResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/events/covers/{coverId}/accept:
    post:
      operationId: acceptEventCover
      summary: Accept a generated cover preview and obtain a reservation id for later event creation
      security:
        - bearerAuth: []
      parameters:
        - name: coverId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptEventCoverResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/events/covers/{coverId}/cancel:
    post:
      operationId: cancelEventCover
      summary: Cancel a generated cover preview (best-effort delete temp object)
      security:
        - bearerAuth: []
      parameters:
        - name: coverId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/events/{id}/cover-url:
    get:
      operationId: getEventCoverUrl
      summary: Get a pre-signed URL to view the event cover image (if configured)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/events:
    get:
      operationId: listEvents
      summary: List events for authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createEvent
      summary: Create event for authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: Resource location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/events/{id}:
    get:
      operationId: getEvent
      summary: Get event by id (only if it belongs to authenticated user)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    GenerateEventCoverRequest:
      type: object
      required: [eventName, categoryLabel, eventTypeLabel, location]
      properties:
        eventName:
          type: string
          minLength: 1
          example: Cena en casa de Pablo
        categoryLabel:
          type: string
          minLength: 1
          example: Food & Drinks
        eventTypeLabel:
          type: string
          minLength: 1
          example: Dinner with Friends
        location:
          type: string
          minLength: 1
          example: Barrancas
        size:
          type: string
          description: Image size; backend may ignore unsupported values
          example: 512x512
    GenerateEventCoverResponse:
      type: object
      required: [coverId, previewUrl, expiresAt]
      properties:
        coverId:
          type: string
          example: 1f0b0c2d-1c8d-4c3b-b7c4-61a5f3c6a0d1
        previewUrl:
          type: string
          example: https://s3.amazonaws.com/... (pre-signed)
        expiresAt:
          type: string
          format: date-time
          example: '2026-02-19T03:00:00Z'
    AcceptEventCoverResponse:
      type: object
      required: [reservationId]
      properties:
        reservationId:
          type: string
          example: 0db0f1ad-9c44-4a2c-9c95-9348d236b0b9
    PresignedUrlResponse:
      type: object
      required: [url, expiresAt]
      properties:
        url:
          type: string
          example: https://s3.amazonaws.com/... (pre-signed)
        expiresAt:
          type: string
          format: date-time
          example: '2026-02-19T03:00:00Z'
    CreateEventRequest:
      type: object
      required: [title, eventTypeId, location, startAt, endAt]
      properties:
        title:
          type: string
          minLength: 1
          example: My event
        eventTypeId:
          type: string
          minLength: 1
          example: birthday
        location:
          type: string
          minLength: 1
          example: San Jose, CR
        startAt:
          type: string
          format: date-time
          example: '2025-01-01T18:00:00Z'
        endAt:
          type: string
          format: date-time
          example: '2025-01-01T22:00:00Z'
        coverReservationId:
          type: string
          nullable: true
          description: Reservation id obtained after accepting an AI-generated cover (optional)
          example: 0db0f1ad-9c44-4a2c-9c95-9348d236b0b9
    Event:
      type: object
      required: [id, ownerId, createdAt, title, eventTypeId, location, startAt, endAt]
      properties:
        id:
          type: string
          example: 7a41e6c0-7a20-4c61-8b2d-8a1df3af7c7b
        ownerId:
          type: string
          example: '109876543210987654321'
        createdAt:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        title:
          type: string
          example: Hello
        eventTypeId:
          type: string
          example: birthday
        location:
          type: string
          example: San Jose, CR
        startAt:
          type: string
          format: date-time
          example: '2025-01-01T18:00:00Z'
        endAt:
          type: string
          format: date-time
          example: '2025-01-01T22:00:00Z'
        coverKey:
          type: string
          nullable: true
          description: S3 key (final bucket) for the event cover PNG
          example: events/7a41e6c0-7a20-4c61-8b2d-8a1df3af7c7b/cover/cover.png
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Event not found